requirements.txt
gunicorn
eventlet
Flask
flask-socketio
flask-sqlalchemy
flask-migrate
flask-marshmallow
marshmallow-sqlalchemy
psycopg2
werkzeug
